+++ apps/api-journeys/src/modules/application/index.ts
@@ -1,10 +1,10 @@
 import { createApplication } from 'graphql-modules'
 import { schemaBuilder } from '@core/shared/util-graphql'
-import block from '../block'
-import journey from '../journey'
-import response from '../response'
+import { blockModule } from '../block'
+import { journeyModule } from '../journey'
+import { responseModule } from '../response'
 
 export default createApplication({
-  modules: [block, journey, response],
+  modules: [blockModule, journeyModule, responseModule],
   schemaBuilder
 })
